does this get us to byte?

+++ prediction.R
@@ -60,7 +60,7 @@ img[img == 0] <- NA
 #is.na(img) <- img[badsegs]
 #img[img %in% badsegs] <- NA
 # Make a vector by replacing segment ids with the prediction
-img.match <- as.numeric(out$pred[match(img, out[,1])])
+img.match <- as.integer(out$pred[match(img, out[,1])])
 # Set the no data value for the output
 img.match[is.na(img.match) == TRUE] <- 0 # Or some other value like -1
 # Set the values of the output raster
